+++ EVCP.MachineLearningModelClient/IMachineLearningModelService.cs
@@ -3,6 +3,6 @@ namespace EVCP.MachineLearningModelClient
 {
     public interface IMachineLearningModelService
     {
-        Task<List<decimal>> Predict(List<ModelInput> modelInputs);
+        Task<List<double>> Predict(List<ModelInput> modelInputs);
     }
 }
